requirements-dev.txt
black==23.7.0
click==8.1.7
oauthlib==3.2.2
pathspec==0.11.2
PyJWT==2.8.0
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
